[#14] feat : 온보딩 페이지 안에서 동작하도록 분리
