+++ app/onboarding/page.css.ts
@@ -9,3 +9,22 @@ export const button = style({
   fontSize: "1.6rem",
   fontWeight: "400",
 });
+
+export const paginationButtons = style({
+  display: "flex",
+  justifyContent: "center",
+  margin: "10px auto 25px ",
+});
+
+export const nextButton = style({
+  width: "10px",
+  height: "10px",
+  border: "1px solid #ccc",
+  borderRadius: "50%",
+  margin: "0 5px",
+  backgroundColor: "#fff",
+});
+
+export const activePaginationButton = style({
+  backgroundColor: "#ccc",
+});
